+++ packages/components/psammead-navigation/src/index.jsx
@@ -11,7 +11,6 @@ import {
 import {
   GEL_GROUP_1_SCREEN_WIDTH_MAX,
   GEL_GROUP_2_SCREEN_WIDTH_MAX,
-  GEL_GROUP_3_SCREEN_WIDTH_MIN,
   GEL_GROUP_5_SCREEN_WIDTH_MIN,
 } from '@bbc/gel-foundations/breakpoints';
 import { getPica } from '@bbc/gel-foundations/typography';
@@ -28,49 +27,10 @@ const BORDER_COLOR = '#eab3b3';
 const C_POSTBOX_TRANSPARENT = `rgba(184, 0, 0, 0)`;
 const C_POSTBOX_OPAQUE = `rgba(184, 0, 0, 1)`;
 
-/* Skip to content */
-const SKIP_LINK_COLOR = '#333';
-const SKIP_LINK_BORDER = '0.1875rem'; // 3px
-const SKIP_LINK_TOP_POSITION_LARGE = '-4rem'; // -64px
-const SKIP_LINK_TOP_POSITION_SMALL = '-3rem'; // -48px
-
 const NavWrapper = styled.div`
   position: relative;
   max-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN};
   margin: 0 auto;
-`;
-
-const SkipLink = styled.a`
-  position: absolute;
-  clip-path: inset(100%);
-  clip: rect(1px, 1px, 1px, 1px);
-  height: 1px;
-  width: 1px;
-  overflow: hidden;
-  padding: ${TOP_BOTTOM_SPACING} ${GEL_SPACING};
-  background-color: #ffffff;
-  border: ${SKIP_LINK_BORDER} solid #000;
-  color: ${SKIP_LINK_COLOR};
-  text-decoration: none;
-  ${({ script }) => script && getPica(script)};
-  ${({ service }) => getSansRegular(service)}
-
-  &:focus {
-    clip-path: none;
-    clip: auto;
-    height: auto;
-    width: auto;
-    top: ${SKIP_LINK_TOP_POSITION_SMALL};
-    left: ${GEL_SPACING};
-
-    @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
-      top: ${SKIP_LINK_TOP_POSITION_LARGE};
-    }
-
-    @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
-      left: 0;
-    }
-  }
 `;
 
 const StyledUnorderedList = styled.ul`
@@ -207,9 +167,9 @@ export const NavigationLi = ({
 };
 
 const StyledNav = styled.nav`
-  background-color: ${C_POSTBOX};
   position: relative;
-
+  background-color: ${C_POSTBOX};
+  border-top: 0.0625rem solid ${C_WHITE};
   ${StyledListItem} {
     ${({ dir }) => css`
       &::after {
@@ -254,7 +214,7 @@ const SwipeableNav = styled.div`
   }
 `;
 
-const Navigation = ({ children, script, skipLinkText, service, dir }) => {
+const Navigation = ({ children, dir }) => {
   const [width, setWidth] = useState(window.innerWidth);
 
   useEffect(() => {
@@ -271,9 +231,6 @@ const Navigation = ({ children, script, skipLinkText, service, dir }) => {
   return (
     <StyledNav role="navigation" dir={dir}>
       <NavWrapper>
-        <SkipLink href="#content" script={script} service={service}>
-          {skipLinkText}
-        </SkipLink>
         <SwipeableNav dir={dir} {...ariaHidden}>
           {React.Children.map(children, child =>
             React.cloneElement(child, { isMobile }),
@@ -286,9 +243,6 @@ const Navigation = ({ children, script, skipLinkText, service, dir }) => {
 
 Navigation.propTypes = {
   children: node.isRequired,
-  script: shape(scriptPropType).isRequired,
-  skipLinkText: string.isRequired,
-  service: string.isRequired,
   dir: oneOf(['ltr', 'rtl']),
 };